Implementação de FaturaCartao

+++ FinancasPessoais/src/main/java/com/financaspessoais/service/BancoService.java
@@ -12,5 +12,5 @@ public interface BancoService {
 
 	List<Banco> listarPorUsuario();
 
-	public Banco buscar(Integer id);
+	public Banco buscar(Short id);
 }
